refactor: tech_events extracted into its own module
