+++ frontend/src/environment.rs
@@ -1,4 +1,3 @@
-use leptos::*;
 #[derive(Debug, Clone)]
 pub struct Environment {
     pub api_url: &'static str,
@@ -6,8 +5,8 @@ pub struct Environment {
 
 pub fn load() -> Environment {
     // dotenv::dotenv().ok();
-    log!("API_URL: {:?}", std::env::var("API_URL"));
     Environment {
+        // api_url: "http://localhost:8000",
         api_url: std::option_env!("API_URL").unwrap_or("https://networking-accumulator.fly.dev/"),
     }
 }
